@@ -1,4 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
+// versions:
+// 	protoc-gen-go v1.36.11
+// 	protoc        v6.33.4
 // source: ml_service.proto
 
 package proto
@@ -8,29 +11,29 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
+	// Verify that this generated code is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
 
 type AnalyzeAudioRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	AudioData     []byte                 `protobuf:"bytes,1,opt,name=audio_data,json=audioData,proto3" json:"audio_data,omitempty"`
+	Format        string                 `protobuf:"bytes,2,opt,name=format,proto3" json:"format,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	AudioData []byte `protobuf:"bytes,1,opt,name=audio_data,json=audioData,proto3" json:"audio_data,omitempty"`
-	Format    string `protobuf:"bytes,2,opt,name=format,proto3" json:"format,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AnalyzeAudioRequest) Reset() {
 	*x = AnalyzeAudioRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ml_service_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ml_service_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AnalyzeAudioRequest) String() string {
@@ -41,7 +44,7 @@ func (*AnalyzeAudioRequest) ProtoMessage() {}
 
 func (x *AnalyzeAudioRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_ml_service_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -51,6 +54,7 @@ func (x *AnalyzeAudioRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
+// Deprecated: Use AnalyzeAudioRequest.ProtoReflect.Descriptor instead.
 func (*AnalyzeAudioRequest) Descriptor() ([]byte, []int) {
 	return file_ml_service_proto_rawDescGZIP(), []int{0}
 }
@@ -70,24 +74,21 @@ func (x *AnalyzeAudioRequest) GetFormat() string {
 }
 
 type AnalyzeAudioResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Embedding     []float32              `protobuf:"fixed32,1,rep,packed,name=embedding,proto3" json:"embedding,omitempty"`
+	MoodEnergy    float32                `protobuf:"fixed32,2,opt,name=mood_energy,json=moodEnergy,proto3" json:"mood_energy,omitempty"`
+	MoodValence   float32                `protobuf:"fixed32,3,opt,name=mood_valence,json=moodValence,proto3" json:"mood_valence,omitempty"`
+	MoodTempo     float32                `protobuf:"fixed32,4,opt,name=mood_tempo,json=moodTempo,proto3" json:"mood_tempo,omitempty"`
+	MoodTexture   float32                `protobuf:"fixed32,5,opt,name=mood_texture,json=moodTexture,proto3" json:"mood_texture,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Embedding   []float32 `protobuf:"fixed32,1,rep,packed,name=embedding,proto3" json:"embedding,omitempty"`
-	MoodEnergy  float32   `protobuf:"fixed32,2,opt,name=mood_energy,json=moodEnergy,proto3" json:"mood_energy,omitempty"`
-	MoodValence float32   `protobuf:"fixed32,3,opt,name=mood_valence,json=moodValence,proto3" json:"mood_valence,omitempty"`
-	MoodTempo   float32   `protobuf:"fixed32,4,opt,name=mood_tempo,json=moodTempo,proto3" json:"mood_tempo,omitempty"`
-	MoodTexture float32   `protobuf:"fixed32,5,opt,name=mood_texture,json=moodTexture,proto3" json:"mood_texture,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AnalyzeAudioResponse) Reset() {
 	*x = AnalyzeAudioResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ml_service_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ml_service_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AnalyzeAudioResponse) String() string {
@@ -98,7 +99,7 @@ func (*AnalyzeAudioResponse) ProtoMessage() {}
 
 func (x *AnalyzeAudioResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_ml_service_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -108,6 +109,7 @@ func (x *AnalyzeAudioResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
+// Deprecated: Use AnalyzeAudioResponse.ProtoReflect.Descriptor instead.
 func (*AnalyzeAudioResponse) Descriptor() ([]byte, []int) {
 	return file_ml_service_proto_rawDescGZIP(), []int{1}
 }
@@ -148,24 +150,21 @@ func (x *AnalyzeAudioResponse) GetMoodTexture() float32 {
 }
 
 type RefineEmbeddingRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	BaseEmbedding []float32              `protobuf:"fixed32,1,rep,packed,name=base_embedding,json=baseEmbedding,proto3" json:"base_embedding,omitempty"`
+	Energy        float32                `protobuf:"fixed32,2,opt,name=energy,proto3" json:"energy,omitempty"`
+	Valence       float32                `protobuf:"fixed32,3,opt,name=valence,proto3" json:"valence,omitempty"`
+	Tempo         float32                `protobuf:"fixed32,4,opt,name=tempo,proto3" json:"tempo,omitempty"`
+	Texture       float32                `protobuf:"fixed32,5,opt,name=texture,proto3" json:"texture,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	BaseEmbedding []float32 `protobuf:"fixed32,1,rep,packed,name=base_embedding,json=baseEmbedding,proto3" json:"base_embedding,omitempty"`
-	Energy        float32   `protobuf:"fixed32,2,opt,name=energy,proto3" json:"energy,omitempty"`
-	Valence       float32   `protobuf:"fixed32,3,opt,name=valence,proto3" json:"valence,omitempty"`
-	Tempo         float32   `protobuf:"fixed32,4,opt,name=tempo,proto3" json:"tempo,omitempty"`
-	Texture       float32   `protobuf:"fixed32,5,opt,name=texture,proto3" json:"texture,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RefineEmbeddingRequest) Reset() {
 	*x = RefineEmbeddingRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ml_service_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ml_service_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RefineEmbeddingRequest) String() string {
@@ -176,7 +175,7 @@ func (*RefineEmbeddingRequest) ProtoMessage() {}
 
 func (x *RefineEmbeddingRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_ml_service_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -186,6 +185,7 @@ func (x *RefineEmbeddingRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
+// Deprecated: Use RefineEmbeddingRequest.ProtoReflect.Descriptor instead.
 func (*RefineEmbeddingRequest) Descriptor() ([]byte, []int) {
 	return file_ml_service_proto_rawDescGZIP(), []int{2}
 }
@@ -226,20 +226,17 @@ func (x *RefineEmbeddingRequest) GetTexture() float32 {
 }
 
 type RefineEmbeddingResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Embedding     []float32              `protobuf:"fixed32,1,rep,packed,name=embedding,proto3" json:"embedding,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Embedding []float32 `protobuf:"fixed32,1,rep,packed,name=embedding,proto3" json:"embedding,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RefineEmbeddingResponse) Reset() {
 	*x = RefineEmbeddingResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ml_service_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ml_service_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RefineEmbeddingResponse) String() string {
@@ -250,7 +247,7 @@ func (*RefineEmbeddingResponse) ProtoMessage() {}
 
 func (x *RefineEmbeddingResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_ml_service_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -260,6 +257,7 @@ func (x *RefineEmbeddingResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
+// Deprecated: Use RefineEmbeddingResponse.ProtoReflect.Descriptor instead.
 func (*RefineEmbeddingResponse) Descriptor() ([]byte, []int) {
 	return file_ml_service_proto_rawDescGZIP(), []int{3}
 }
@@ -272,18 +270,16 @@ func (x *RefineEmbeddingResponse) GetEmbedding() []float32 {
 }
 
 type HealthCheckRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *HealthCheckRequest) Reset() {
 	*x = HealthCheckRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ml_service_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ml_service_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *HealthCheckRequest) String() string {
@@ -294,7 +290,7 @@ func (*HealthCheckRequest) ProtoMessage() {}
 
 func (x *HealthCheckRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_ml_service_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -304,26 +300,24 @@ func (x *HealthCheckRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
+// Deprecated: Use HealthCheckRequest.ProtoReflect.Descriptor instead.
 func (*HealthCheckRequest) Descriptor() ([]byte, []int) {
 	return file_ml_service_proto_rawDescGZIP(), []int{4}
 }
 
 type HealthCheckResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Healthy       bool                   `protobuf:"varint,1,opt,name=healthy,proto3" json:"healthy,omitempty"`
+	Message       string                 `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Healthy bool   `protobuf:"varint,1,opt,name=healthy,proto3" json:"healthy,omitempty"`
-	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *HealthCheckResponse) Reset() {
 	*x = HealthCheckResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ml_service_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ml_service_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *HealthCheckResponse) String() string {
@@ -334,7 +328,7 @@ func (*HealthCheckResponse) ProtoMessage() {}
 
 func (x *HealthCheckResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_ml_service_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -344,6 +338,7 @@ func (x *HealthCheckResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
+// Deprecated: Use HealthCheckResponse.ProtoReflect.Descriptor instead.
 func (*HealthCheckResponse) Descriptor() ([]byte, []int) {
 	return file_ml_service_proto_rawDescGZIP(), []int{5}
 }
@@ -364,79 +359,52 @@ func (x *HealthCheckResponse) GetMessage() string {
 
 var File_ml_service_proto protoreflect.FileDescriptor
 
-var file_ml_service_proto_rawDesc = []byte{
-	0x0a, 0x10, 0x6d, 0x6c, 0x5f, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x12, 0x05, 0x65, 0x76, 0x6f, 0x6b, 0x65, 0x22, 0x4a, 0x0a, 0x13, 0x41, 0x6e, 0x61,
-	0x6c, 0x79, 0x7a, 0x65, 0x41, 0x75, 0x64, 0x69, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x12, 0x1d, 0x0a, 0x0a, 0x61, 0x75, 0x64, 0x69, 0x6f, 0x5f, 0x64, 0x61, 0x74, 0x61, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x09, 0x61, 0x75, 0x64, 0x69, 0x6f, 0x44, 0x61, 0x74, 0x61, 0x12,
-	0x16, 0x0a, 0x06, 0x66, 0x6f, 0x72, 0x6d, 0x61, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x06, 0x66, 0x6f, 0x72, 0x6d, 0x61, 0x74, 0x22, 0xae, 0x01, 0x0a, 0x14, 0x41, 0x6e, 0x61, 0x6c,
-	0x79, 0x7a, 0x65, 0x41, 0x75, 0x64, 0x69, 0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
-	0x12, 0x1c, 0x0a, 0x09, 0x65, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e, 0x67, 0x18, 0x01, 0x20,
-	0x03, 0x28, 0x02, 0x52, 0x09, 0x65, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e, 0x67, 0x12, 0x1f,
-	0x0a, 0x0b, 0x6d, 0x6f, 0x6f, 0x64, 0x5f, 0x65, 0x6e, 0x65, 0x72, 0x67, 0x79, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x02, 0x52, 0x0a, 0x6d, 0x6f, 0x6f, 0x64, 0x45, 0x6e, 0x65, 0x72, 0x67, 0x79, 0x12,
-	0x21, 0x0a, 0x0c, 0x6d, 0x6f, 0x6f, 0x64, 0x5f, 0x76, 0x61, 0x6c, 0x65, 0x6e, 0x63, 0x65, 0x18,
-	0x03, 0x20, 0x01, 0x28, 0x02, 0x52, 0x0b, 0x6d, 0x6f, 0x6f, 0x64, 0x56, 0x61, 0x6c, 0x65, 0x6e,
-	0x63, 0x65, 0x12, 0x1d, 0x0a, 0x0a, 0x6d, 0x6f, 0x6f, 0x64, 0x5f, 0x74, 0x65, 0x6d, 0x70, 0x6f,
-	0x18, 0x04, 0x20, 0x01, 0x28, 0x02, 0x52, 0x09, 0x6d, 0x6f, 0x6f, 0x64, 0x54, 0x65, 0x6d, 0x70,
-	0x6f, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x6f, 0x6f, 0x64, 0x5f, 0x74, 0x65, 0x78, 0x74, 0x75, 0x72,
-	0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x02, 0x52, 0x0b, 0x6d, 0x6f, 0x6f, 0x64, 0x54, 0x65, 0x78,
-	0x74, 0x75, 0x72, 0x65, 0x22, 0x95, 0x01, 0x0a, 0x16, 0x52, 0x65, 0x66, 0x69, 0x6e, 0x65, 0x45,
-	0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e, 0x67, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
-	0x25, 0x0a, 0x0e, 0x62, 0x61, 0x73, 0x65, 0x5f, 0x65, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e,
-	0x67, 0x18, 0x01, 0x20, 0x03, 0x28, 0x02, 0x52, 0x0d, 0x62, 0x61, 0x73, 0x65, 0x45, 0x6d, 0x62,
-	0x65, 0x64, 0x64, 0x69, 0x6e, 0x67, 0x12, 0x16, 0x0a, 0x06, 0x65, 0x6e, 0x65, 0x72, 0x67, 0x79,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x02, 0x52, 0x06, 0x65, 0x6e, 0x65, 0x72, 0x67, 0x79, 0x12, 0x18,
-	0x0a, 0x07, 0x76, 0x61, 0x6c, 0x65, 0x6e, 0x63, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x02, 0x52,
-	0x07, 0x76, 0x61, 0x6c, 0x65, 0x6e, 0x63, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x74, 0x65, 0x6d, 0x70,
-	0x6f, 0x18, 0x04, 0x20, 0x01, 0x28, 0x02, 0x52, 0x05, 0x74, 0x65, 0x6d, 0x70, 0x6f, 0x12, 0x18,
-	0x0a, 0x07, 0x74, 0x65, 0x78, 0x74, 0x75, 0x72, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x02, 0x52,
-	0x07, 0x74, 0x65, 0x78, 0x74, 0x75, 0x72, 0x65, 0x22, 0x37, 0x0a, 0x17, 0x52, 0x65, 0x66, 0x69,
-	0x6e, 0x65, 0x45, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e, 0x67, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x1c, 0x0a, 0x09, 0x65, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e, 0x67,
-	0x18, 0x01, 0x20, 0x03, 0x28, 0x02, 0x52, 0x09, 0x65, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e,
-	0x67, 0x22, 0x14, 0x0a, 0x12, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x43, 0x68, 0x65, 0x63, 0x6b,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x49, 0x0a, 0x13, 0x48, 0x65, 0x61, 0x6c, 0x74,
-	0x68, 0x43, 0x68, 0x65, 0x63, 0x6b, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x18,
-	0x0a, 0x07, 0x68, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52,
-	0x07, 0x68, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x79, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73,
-	0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x32, 0xcb, 0x01, 0x0a, 0x09, 0x4d, 0x4c, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65,
-	0x12, 0x43, 0x0a, 0x0c, 0x41, 0x6e, 0x61, 0x6c, 0x79, 0x7a, 0x65, 0x41, 0x75, 0x64, 0x69, 0x6f,
-	0x12, 0x1a, 0x2e, 0x65, 0x76, 0x6f, 0x6b, 0x65, 0x2e, 0x41, 0x6e, 0x61, 0x6c, 0x79, 0x7a, 0x65,
-	0x41, 0x75, 0x64, 0x69, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1b, 0x2e, 0x65,
-	0x76, 0x6f, 0x6b, 0x65, 0x2e, 0x41, 0x6e, 0x61, 0x6c, 0x79, 0x7a, 0x65, 0x41, 0x75, 0x64, 0x69,
-	0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x4c, 0x0a, 0x0f, 0x52, 0x65, 0x66,
-	0x69, 0x6e, 0x65, 0x45, 0x6d, 0x62, 0x65, 0x64, 0x64, 0x69, 0x6e, 0x67, 0x12, 0x1d, 0x2e, 0x65,
-	0x76, 0x6f, 0x6b, 0x65, 0x2e, 0x52, 0x65, 0x66, 0x69, 0x6e, 0x65, 0x45, 0x6d, 0x62, 0x65, 0x64,
-	0x64, 0x69, 0x6e, 0x67, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x65, 0x76,
-	0x6f, 0x6b, 0x65, 0x2e, 0x52, 0x65, 0x66, 0x69, 0x6e, 0x65, 0x45, 0x6d, 0x62, 0x65, 0x64, 0x64,
-	0x69, 0x6e, 0x67, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x40, 0x0a, 0x0b, 0x48,
-	0x65, 0x61, 0x6c, 0x74, 0x68, 0x43, 0x68, 0x65, 0x63, 0x6b, 0x12, 0x19, 0x2e, 0x65, 0x76, 0x6f,
-	0x6b, 0x65, 0x2e, 0x48, 0x65, 0x61, 0x6c, 0x74, 0x68, 0x43, 0x68, 0x65, 0x63, 0x6b, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1a, 0x2e, 0x65, 0x76, 0x6f, 0x6b, 0x65, 0x2e, 0x48, 0x65,
-	0x61, 0x6c, 0x74, 0x68, 0x43, 0x68, 0x65, 0x63, 0x6b, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
-	0x65, 0x42, 0x22, 0x5a, 0x20, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f,
-	0x65, 0x76, 0x6f, 0x6b, 0x65, 0x2f, 0x62, 0x61, 0x63, 0x6b, 0x65, 0x6e, 0x64, 0x2f, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_ml_service_proto_rawDesc = "" +
+	"\n" +
+	"\x10ml_service.proto\x12\x05evoke\"L\n" +
+	"\x13AnalyzeAudioRequest\x12\x1d\n" +
+	"\n" +
+	"audio_data\x18\x01 \x01(\fR\taudioData\x12\x16\n" +
+	"\x06format\x18\x02 \x01(\tR\x06format\"\xba\x01\n" +
+	"\x14AnalyzeAudioResponse\x12\x1c\n" +
+	"\tembedding\x18\x01 \x03(\x02R\tembedding\x12\x1f\n" +
+	"\vmood_energy\x18\x02 \x01(\x02R\n" +
+	"moodEnergy\x12!\n" +
+	"\fmood_valence\x18\x03 \x01(\x02R\vmoodValence\x12\x1d\n" +
+	"\n" +
+	"mood_tempo\x18\x04 \x01(\x02R\tmoodTempo\x12!\n" +
+	"\fmood_texture\x18\x05 \x01(\x02R\vmoodTexture\"\xa1\x01\n" +
+	"\x16RefineEmbeddingRequest\x12%\n" +
+	"\x0ebase_embedding\x18\x01 \x03(\x02R\rbaseEmbedding\x12\x16\n" +
+	"\x06energy\x18\x02 \x01(\x02R\x06energy\x12\x18\n" +
+	"\avalence\x18\x03 \x01(\x02R\avalence\x12\x14\n" +
+	"\x05tempo\x18\x04 \x01(\x02R\x05tempo\x12\x18\n" +
+	"\atexture\x18\x05 \x01(\x02R\atexture\"7\n" +
+	"\x17RefineEmbeddingResponse\x12\x1c\n" +
+	"\tembedding\x18\x01 \x03(\x02R\tembedding\"\x14\n" +
+	"\x12HealthCheckRequest\"I\n" +
+	"\x13HealthCheckResponse\x12\x18\n" +
+	"\ahealthy\x18\x01 \x01(\bR\ahealthy\x12\x18\n" +
+	"\amessage\x18\x02 \x01(\tR\amessage2\xec\x01\n" +
+	"\tMLService\x12G\n" +
+	"\fAnalyzeAudio\x12\x1a.evoke.AnalyzeAudioRequest\x1a\x1b.evoke.AnalyzeAudioResponse\x12P\n" +
+	"\x0fRefineEmbedding\x12\x1d.evoke.RefineEmbeddingRequest\x1a\x1e.evoke.RefineEmbeddingResponse\x12D\n" +
+	"\vHealthCheck\x12\x19.evoke.HealthCheckRequest\x1a\x1a.evoke.HealthCheckResponseB Z\x1egithub.com/evoke/backend/protob\x06proto3"
 
 var (
 	file_ml_service_proto_rawDescOnce sync.Once
-	file_ml_service_proto_rawDescData = file_ml_service_proto_rawDesc
+	file_ml_service_proto_rawDescData []byte
 )
 
 func file_ml_service_proto_rawDescGZIP() []byte {
 	file_ml_service_proto_rawDescOnce.Do(func() {
-		file_ml_service_proto_rawDescData = protoimpl.X.CompressGZIP(file_ml_service_proto_rawDescData)
+		file_ml_service_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_ml_service_proto_rawDesc), len(file_ml_service_proto_rawDesc)))
 	})
 	return file_ml_service_proto_rawDescData
 }
 
 var file_ml_service_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_ml_service_proto_goTypes = []interface{}{
+var file_ml_service_proto_goTypes = []any{
 	(*AnalyzeAudioRequest)(nil),     // 0: evoke.AnalyzeAudioRequest
 	(*AnalyzeAudioResponse)(nil),    // 1: evoke.AnalyzeAudioResponse
 	(*RefineEmbeddingRequest)(nil),  // 2: evoke.RefineEmbeddingRequest
@@ -459,90 +427,15 @@ var file_ml_service_proto_depIdxs = []int32{
 }
 
 func init() { file_ml_service_proto_init() }
-
 func file_ml_service_proto_init() {
 	if File_ml_service_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_ml_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*AnalyzeAudioRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_ml_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*AnalyzeAudioResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_ml_service_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RefineEmbeddingRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_ml_service_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RefineEmbeddingResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_ml_service_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*HealthCheckRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_ml_service_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*HealthCheckResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_ml_service_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_ml_service_proto_rawDesc), len(file_ml_service_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   6,
 			NumExtensions: 0,
@@ -553,7 +446,6 @@ func file_ml_service_proto_init() {
 		MessageInfos:      file_ml_service_proto_msgTypes,
 	}.Build()
 	File_ml_service_proto = out.File
-	file_ml_service_proto_rawDescGZIP()
 	file_ml_service_proto_goTypes = nil
 	file_ml_service_proto_depIdxs = nil
 }
