@@ -1,4 +1,8 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
+// versions:
+// - protoc-gen-go-grpc v1.6.1
+// - protoc             v6.33.4
+// source: ml_service.proto
 
 package proto
 
@@ -11,7 +15,14 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	MLService_AnalyzeAudio_FullMethodName    = "/evoke.MLService/AnalyzeAudio"
+	MLService_RefineEmbedding_FullMethodName = "/evoke.MLService/RefineEmbedding"
+	MLService_HealthCheck_FullMethodName     = "/evoke.MLService/HealthCheck"
+)
 
 // MLServiceClient is the client API for MLService service.
 //
@@ -31,8 +42,9 @@ func NewMLServiceClient(cc grpc.ClientConnInterface) MLServiceClient {
 }
 
 func (c *mLServiceClient) AnalyzeAudio(ctx context.Context, in *AnalyzeAudioRequest, opts ...grpc.CallOption) (*AnalyzeAudioResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AnalyzeAudioResponse)
-	err := c.cc.Invoke(ctx, "/evoke.MLService/AnalyzeAudio", in, out, opts...)
+	err := c.cc.Invoke(ctx, MLService_AnalyzeAudio_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -40,8 +52,9 @@ func (c *mLServiceClient) AnalyzeAudio(ctx context.Context, in *AnalyzeAudioRequ
 }
 
 func (c *mLServiceClient) RefineEmbedding(ctx context.Context, in *RefineEmbeddingRequest, opts ...grpc.CallOption) (*RefineEmbeddingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(RefineEmbeddingResponse)
-	err := c.cc.Invoke(ctx, "/evoke.MLService/RefineEmbedding", in, out, opts...)
+	err := c.cc.Invoke(ctx, MLService_RefineEmbedding_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -49,8 +62,9 @@ func (c *mLServiceClient) RefineEmbedding(ctx context.Context, in *RefineEmbeddi
 }
 
 func (c *mLServiceClient) HealthCheck(ctx context.Context, in *HealthCheckRequest, opts ...grpc.CallOption) (*HealthCheckResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(HealthCheckResponse)
-	err := c.cc.Invoke(ctx, "/evoke.MLService/HealthCheck", in, out, opts...)
+	err := c.cc.Invoke(ctx, MLService_HealthCheck_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,7 +73,7 @@ func (c *mLServiceClient) HealthCheck(ctx context.Context, in *HealthCheckReques
 
 // MLServiceServer is the server API for MLService service.
 // All implementations must embed UnimplementedMLServiceServer
-// for forward compatibility
+// for forward compatibility.
 type MLServiceServer interface {
 	AnalyzeAudio(context.Context, *AnalyzeAudioRequest) (*AnalyzeAudioResponse, error)
 	RefineEmbedding(context.Context, *RefineEmbeddingRequest) (*RefineEmbeddingResponse, error)
@@ -67,20 +81,24 @@ type MLServiceServer interface {
 	mustEmbedUnimplementedMLServiceServer()
 }
 
-// UnimplementedMLServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedMLServiceServer struct {
-}
+// UnimplementedMLServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMLServiceServer struct{}
 
 func (UnimplementedMLServiceServer) AnalyzeAudio(context.Context, *AnalyzeAudioRequest) (*AnalyzeAudioResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method AnalyzeAudio not implemented")
+	return nil, status.Error(codes.Unimplemented, "method AnalyzeAudio not implemented")
 }
 func (UnimplementedMLServiceServer) RefineEmbedding(context.Context, *RefineEmbeddingRequest) (*RefineEmbeddingResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method RefineEmbedding not implemented")
+	return nil, status.Error(codes.Unimplemented, "method RefineEmbedding not implemented")
 }
 func (UnimplementedMLServiceServer) HealthCheck(context.Context, *HealthCheckRequest) (*HealthCheckResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method HealthCheck not implemented")
+	return nil, status.Error(codes.Unimplemented, "method HealthCheck not implemented")
 }
 func (UnimplementedMLServiceServer) mustEmbedUnimplementedMLServiceServer() {}
+func (UnimplementedMLServiceServer) testEmbeddedByValue()                   {}
 
 // UnsafeMLServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MLServiceServer will
@@ -90,6 +108,13 @@ type UnsafeMLServiceServer interface {
 }
 
 func RegisterMLServiceServer(s grpc.ServiceRegistrar, srv MLServiceServer) {
+	// If the following call panics, it indicates UnimplementedMLServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MLService_ServiceDesc, srv)
 }
 
@@ -103,7 +128,7 @@ func _MLService_AnalyzeAudio_Handler(srv interface{}, ctx context.Context, dec f
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/evoke.MLService/AnalyzeAudio",
+		FullMethod: MLService_AnalyzeAudio_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(MLServiceServer).AnalyzeAudio(ctx, req.(*AnalyzeAudioRequest))
@@ -121,7 +146,7 @@ func _MLService_RefineEmbedding_Handler(srv interface{}, ctx context.Context, de
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/evoke.MLService/RefineEmbedding",
+		FullMethod: MLService_RefineEmbedding_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(MLServiceServer).RefineEmbedding(ctx, req.(*RefineEmbeddingRequest))
@@ -139,7 +164,7 @@ func _MLService_HealthCheck_Handler(srv interface{}, ctx context.Context, dec fu
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/evoke.MLService/HealthCheck",
+		FullMethod: MLService_HealthCheck_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(MLServiceServer).HealthCheck(ctx, req.(*HealthCheckRequest))
